+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix find-references and references lens for `defrecord`.
 - Improve `clean-ns` refactoring following the Clojure style guide.
 - Deprecate `:keep-require-at-start?` in favor of `:clean :ns-inner-blocks-indentation` with default value of `:next-line`.
+- Bump `org.xerial/sqlite-jdbc` from 3.34.0 to 3.36.0.
 
 ## 2021.06.24-14.24.11