+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix `clean-ns` corner case issue where it would replace the `:as` value with the `:refer` value.
 - Deprecate `:keep-require-at-start?` in favor of `:clean :ns-inner-blocks-indentation` with default value of `:next-line`.
 - Bump `org.xerial/sqlite-jdbc` from 3.34.0 to 3.36.0.
+- Avoid adding duplicated namespace for new blank files.
 
 ## 2021.06.24-14.24.11