+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Deprecate `:keep-require-at-start?` in favor of `:clean :ns-inner-blocks-indentation` with default value of `:next-line`.
 - Bump `org.xerial/sqlite-jdbc` from 3.34.0 to 3.36.0.
 - Avoid adding duplicated namespace for new blank files.
+- Add API / CLI support, form more information check the [API](https://clojure-lsp.github.io/clojure-lsp/api/) section.
 
 ## 2021.06.24-14.24.11