+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2021.07.01-13.46.18
+
 - Fix find-references and references lens for `defrecord`.
 - Improve `clean-ns` refactoring following the Clojure style guide.
 - Fix `clean-ns` corner case issue where it would replace the `:as` value with the `:refer` value.